[Feat] Character Touch Interaction
캐릭터 터치 시 점프하는 상호작용 기능을 구현했습니다.

+++ Assets/Scripts/CharacterModel.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class CharacterModel : MonoBehaviour
 {
-    [SerializeField] int fullness;
-    public int Fullness { get { return fullness; } set { fullness = value; } }
+    [SerializeField] float fullness;
+    public float Fullness { get { return fullness; } set { fullness = value; } }
 
     [SerializeField] int interactCount;
     public int InteractCount { get { return interactCount; } set { interactCount = value; } }
